+++ bhp/lims/config.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # Copyright 2018 Botswana Harvard Partnership (BHP)
-
+#
 from Products.Archetypes.utils import DisplayList
 
 PRODUCT_NAME = "bhp.lims"
@@ -15,4 +15,10 @@ GENDERS_ANY = DisplayList((
     ('a', "Any"),
     ('m', "Male"),
     ('f', "Female"),
+))
+
+PRIORITIES = DisplayList((
+    ('1', 'Urgent'),
+    ('3', 'Routine'),
+    ('5', 'STAT'),
 ))